Make sure to only use specified VERSION_PREFIX
If it happens that you have tags in your repo mixing plain semantic versioning with versions with a prefix, current line could match the wrong tag. For instance if you have tags in your repo:

```
1.5.0
v1.5.0
```
and `export VERSION_PREFIX=` you'll get tag `v1.5.0`.

This fix makes sure that the regular expression starts at the beginning of the line so making the `VERSION_PREFIX` a real filter.